Make it possible for multiple games to connect using the same room name.  All new connections are stored together, and after each connection an attempt is made to build one complete game with a host and however many guests.  All remaining devices are moved into a new pending record in the same state, and the completed game is treated as always.  Seems to work, though nearly 20% of linux instances are failing to connect the relay run from the new test script samename.sh. Need to figure out why.
Also added logging of seed and connname to comms.c since games
launched together can no longer be certain to connect on the relay.
This allows the test script to identify joined games from their logs
and detect success or failure.

This checkin changes the relay protocol, so relay and clients will
both need to be upgraded.