This comment meant for this fix, not prev! :
Fix board sometimes coming up half-height by indicating that it should
never have a soft keyboard associated with it.  It's otherwise getting
told its height is half what it should be, and I can't find any way to
get the layout redone when the keyboard goes away.  (The keyboard in
this case is leftover from the previous activity.)